# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "adminpanel",
    "description": "Service for admin panel & mobile AP now",
    "version": "1.0"
  },
  "produces": ["application/json"],
  "paths": {
...
    "/api/file": {
      "get": {
        "tags": [
          "file"
        ],
        "summary": "List all files",
        "operationId": "getfiles",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/file"
                      }
                    },
                    "message": {
                      "type": "string",
                      "default": "Files found successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "User not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "403": {
            "description": "You are not logged in!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error get files",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "file"
        ],
        "summary": "Create file",
        "operationId": "postfile",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "userId": {
                    "type": "string",
                    "example": "61b85603a77df173ac4df406"
                  },
                  "file": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/file"
                    },
                    "message": {
                      "type": "string",
                      "default": "File created successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "User not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "403": {
            "description": "You are not logged in!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiResponse"
                }
              }
            }
          },
          "default": {
            "description": "File not created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/field": {
      "get": {
        "tags": [
          "field"
        ],
        "summary": "List all fields",
        "operationId": "getfields",
        "responses": {
          "200": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/field"
                      }
                    },
                    "message": {
                      "type": "string",
                      "default": "Fields found successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "User not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "403": {
            "description": "You are not logged in!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiResponse"
                }
              }
            }
          },
          "default": {
            "description": "Error get fields",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "field"
        ],
        "summary": "Create field",
        "operationId": "postfield",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/field"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "successful operation",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "default": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/field"
                    },
                    "message": {
                      "type": "string",
                      "default": "Field created successfully"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "User not exist",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          },
          "403": {
            "description": "You are not logged in!",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/apiResponse"
                }
              }
            }
          },
          "default": {
            "description": "Field not created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/errorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
      "schemas": {
...
        "file": {
          "required": [
            "userId",
            "originalname",
            "size",
            "destination",
            "filename"
          ],
          "type": "object",
          "properties": {
            "_id": {
              "type": "string"
            },
            "userId": {
              "type": "string"
            },
            "originalname": {
              "type": "string"
            },
            "encoding": {
              "type": "string"
            },
            "mimetype": {
              "type": "string"
            },
            "size": {
              "type": "integer",
              "format": "int32"
            },
            "destination": {
              "type": "string"
            },
            "filename": {
              "type": "string"
            },
            "createdAt": {
              "type": "string",
              "format": "date-time"
            },
            "updatedAt": {
              "type": "string",
              "format": "date-time"
            }
          },
          "xml": {
            "name": "file"
          }
        },
        "field": {
          "required": [
            "type"
          ],
          "type": "object",
          "properties": {
            "_id": {
              "type": "string"
            },
            "type": {
              "type": "string"
            },
            "arrayString": {
              "type": "array",
              "items": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              }
            },
            "arrayNumber": {
              "type": "array",
              "items": {
                "type": "number",
                "format": "float"
              }
            },
            "createdAt": {
              "type": "string",
              "format": "date-time"
            },
            "updatedAt": {
              "type": "string",
              "format": "date-time"
            }
          },
          "xml": {
            "name": "field"
          }
        },
        "errorResponse": {
          "type": "object",
          "properties": {
            "success": {
              "type": "boolean",
              "default": false
            },
            "status": {
              "type": "string",
              "default": "error"
            },
            "message": {
              "type": "string",
              "default": "Not Found!"
            }
          },
          "xml": {
            "name": "errorResponse"
          }
        },
        "apiResponse": {
          "type": "object",
          "properties": {
            "success": {
              "type": "boolean",
              "default": true
            },
            "message": {
              "type": "string"
            }
          },
          "xml": {
            "name": "errorResponse"
          }
        }
      }
    }
 }